add component refactoring
